requirements.txt
streamlit>=1.28.0
openai>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.31.0
aiohttp>=3.9.0
google-generativeai>=0.3.0
tenacity>=8.2.0
